Add files via upload
The model is a hybrid of a CNN (convolutional neural network - VGG16 and machine learning model - XGBoost. The VGG16 consist of 16 layers that include 13 convolutional layers and 3 fully connected layers. It uses a 3x3 kernel for convolution and max pooling layers to rogressively reduce the image imension while extracting heirarchial features. The activation function applied for each layer is ReLU and this architecture is widely used for image recognition tasks due to its strong features extraction capabilities. the XGBoost algorithm is based on gradient boosting and builds an ensemble of decision trees which helps in appropriately and reliably classify the images.